# swagger.yaml
openapi: 3.0.0
info:
  title: Task Management API
  version: 1.0.0
  description: API for managing users, projects, tasks, and comments
servers:
  - url: https://final-77ir.onrender.com
    description: Production server
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Get all users (Admin only)
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of users
        401:
          description: Unauthorized
        403:
          description: Forbidden

    post:
      tags: [Users]
      summary: Create user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Jane Doe
              email: jane@example.com
              password: Password123!
              role: USER
      responses:
        201:
          description: User created
        400:
          description: Invalid request body
        401:
          description: Unauthorized
        403:
          description: Forbidden
        409:
          description: Email already exists
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: User found
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: User not found

    put:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Updated Name
              email: updated@example.com
              role: USER
      responses:
        200:
          description: User updated
        400:
          description: Invalid request body
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: User not found

    delete:
      tags: [Users]
      summary: Delete user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: User deleted
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: User not found
...